Move command classes
