+++ src/main/java/decodex/parser/RecipeCommandParser.java
@@ -5,10 +5,10 @@ import decodex.logic.commands.recipe.RecipeDeleteCommand;
 import decodex.commands.recipe.RecipeDeselectCommand;
 import decodex.logic.commands.recipe.RecipeListCommand;
 import decodex.logic.commands.recipe.RecipeNewCommand;
-import decodex.commands.recipe.RecipePopCommand;
-import decodex.commands.recipe.RecipePushCommand;
-import decodex.commands.recipe.RecipeResetCommand;
-import decodex.commands.recipe.RecipeSelectCommand;
+import decodex.logic.commands.recipe.RecipePopCommand;
+import decodex.logic.commands.recipe.RecipePushCommand;
+import decodex.logic.commands.recipe.RecipeResetCommand;
+import decodex.logic.commands.recipe.RecipeSelectCommand;
 import decodex.data.exception.CommandException;
 import decodex.ui.messages.ErrorMessages;
 import java.util.Arrays;
